feat(booking): prefill what the browser and the device already know
The booking form asked repeat visitors to retype their own name and email
every time, and gave the browser nothing to autofill from.

Three tiers, in order of how much they assume. Every field now carries the
autofill token that matches it, which costs nothing and covers most of the
value. The visitor's OWN contact details are remembered on this device
only, with a visible "Not you? Clear" for shared machines and a line saying
where they came from — never the address or services, which belong to one
property rather than one person. The signed-in agent's identity needs no
field at all: the hold takes it from the session.

An agent's booking is deliberately excluded from both remembering and
browser autofill: those fields hold their client's details, so prefilling
them from this device would put one client's name on another's booking.

ScheduleStep moves to its own module — BookingSteps crossed the file-size
cap, and the step boundary is the natural seam.
