+++ app/routes/public/booking.test.tsx
@@ -6,8 +6,8 @@
  *
  * An anonymous visitor must see exactly what they saw before.
  */
-import { describe, it, expect } from "vitest";
-import { render, fireEvent, waitFor } from "@testing-library/react";
+import { describe, it, expect, beforeEach } from "vitest";
+import { render, fireEvent, waitFor, screen } from "@testing-library/react";
 import { createRoutesStub } from "react-router";
 
 import BookingPage from "~/routes/public/booking";
@@ -56,6 +56,61 @@ describe("BookingPage — anonymous visitor", () => {
   });
 });
 
+describe("BookingPage — prefill", () => {
+  beforeEach(() => localStorage.clear());
+
+  it("marks every field with the browser autofill token that matches it", async () => {
+    const { findByPlaceholderText, container } = renderBooking({ agentBooking: null });
+    const address = await findByPlaceholderText(/123 Main St/i);
+    expect(address.getAttribute("autocomplete")).toBe("street-address");
+    fireEvent.change(address, { target: { value: "123 Main St" } });
+    fireEvent.click(await screen.findByText("Continue"));
+    fireEvent.click(await screen.findByText("Full Inspection"));
+    fireEvent.click(await screen.findByText("Continue"));
+    expect(container.querySelector("input[type='email']")?.getAttribute("autocomplete")).toBe("email");
+    expect(screen.getByPlaceholderText("Jane Doe").getAttribute("autocomplete")).toBe("name");
+  });
+
+  it("remembers the visitor's own contact details and offers to clear them", async () => {
+    localStorage.setItem(
+      "oi.booking.contact",
+      JSON.stringify({ name: "Sarah Buyer", email: "sarah@example.com" }),
+    );
+    const { findByPlaceholderText } = renderBooking({ agentBooking: null });
+    fireEvent.change(await findByPlaceholderText(/123 Main St/i), { target: { value: "1 A St" } });
+    fireEvent.click(await screen.findByText("Continue"));
+    fireEvent.click(await screen.findByText("Full Inspection"));
+    fireEvent.click(await screen.findByText("Continue"));
+
+    const name = screen.getByPlaceholderText("Jane Doe") as HTMLInputElement;
+    expect(name.value).toBe("Sarah Buyer");
+    // Says it is stored on this device, and lets a different visitor drop it.
+    expect(screen.getByText(/this browser|this device/i)).toBeTruthy();
+    fireEvent.click(screen.getByText(/not you/i));
+    expect((screen.getByPlaceholderText("Jane Doe") as HTMLInputElement).value).toBe("");
+    expect(localStorage.getItem("oi.booking.contact")).toBeNull();
+  });
+
+  it("never prefills a remembered contact into an agent's booking — that would be someone else's client", async () => {
+    localStorage.setItem(
+      "oi.booking.contact",
+      JSON.stringify({ name: "Sarah Buyer", email: "sarah@example.com" }),
+    );
+    const { findByPlaceholderText } = renderBooking({
+      agentBooking: { agentName: "Jane Smith", tenantId: "t-1" },
+    });
+    fireEvent.change(await findByPlaceholderText(/123 Main St/i), { target: { value: "1 A St" } });
+    fireEvent.click(await screen.findByText("Continue"));
+    fireEvent.click(await screen.findByText("Full Inspection"));
+    fireEvent.click(await screen.findByText("Continue"));
+
+    const name = screen.getByPlaceholderText("Jane Doe") as HTMLInputElement;
+    expect(name.value).toBe("");
+    // The fields are the client's, so the browser's own saved identity is wrong here too.
+    expect(name.getAttribute("autocomplete")).toBe("off");
+  });
+});
+
 describe("BookingPage — signed-in agent", () => {
   it("says whose behalf the booking is on", async () => {
     const { findByText } = renderBooking({
@@ -97,5 +152,11 @@ describe("BookingPage — signed-in agent", () => {
     expect(form.get("clientName")).toBe("Sarah Buyer");
     expect(form.get("clientEmail")).toBe("sarah@example.com");
     expect(form.getAll("serviceId")).toEqual(["svc-1"]);
+    // Tier 3 of prefill: the agent never types who they are. Their identity —
+    // and therefore the referral credit — comes from the session on the server,
+    // so nothing in the submission can be forged to claim someone else's.
+    expect(form.get("agentRefSlug")).toBeNull();
+    expect(form.get("agentUserId")).toBeNull();
+    expect(form.get("tenantId")).toBeNull();
   });
 });
